Eliminate several casts to signed integers
Do this by changing the variable types to signed integers.

+++ rc4hash.c
@@ -1,7 +1,7 @@
 /*
  * rc4hash - abuse ARCFOUR as a simplistic and reasonably fast hash algorithm
  *
- * Version 2020.322.3
+ * Version 2020.322.4
  *
  * Copyright (c) 2020 Guenther Brunthaler. All rights reserved.
  *
@@ -38,7 +38,7 @@ int main(int argc, char **argv) {
    int a= 0;
    long digest_bits= 256;
    char const *alphabet= b32custom_alphabet;
-   unsigned alphabet_bitmask= DIM(b32custom_alphabet) - 1, alphabet_bits;
+   int alphabet_bitmask= (int)DIM(b32custom_alphabet) - 1, alphabet_bits;
    ARCFOUR_VARDEFS(static);
    {
       int optpos= 0;
@@ -47,7 +47,8 @@ int main(int argc, char **argv) {
          switch (opt= getopt_simplest(&a, &optpos, argc, argv)) {
             case 0: goto no_more_options;
             case 'x':
-               alphabet= hex_alphabet; alphabet_bitmask= DIM(hex_alphabet) - 1;
+               alphabet= hex_alphabet;
+               alphabet_bitmask= (int)DIM(hex_alphabet) - 1;
                break;
             case 'r': alphabet= 0; alphabet_bitmask= (1 << 8) - 1; break;
             case 'B': case 'b':
@@ -71,12 +72,12 @@ int main(int argc, char **argv) {
    }
    no_more_options:
    {
-      unsigned bm;
+      int bm;
       for (alphabet_bits= bm= 0; bm != alphabet_bitmask; ++alphabet_bits) {
          bm+= bm + 1;
       }
    }
-   if (digest_bits & (long)(alphabet_bits - 1)) {
+   if (digest_bits & alphabet_bits - 1) {
       bad_digest_size:
       error= "Invalid digest size requested!"; goto fail;
    }
@@ -125,11 +126,11 @@ int main(int argc, char **argv) {
       /* Produce the message digest. */
       {
          long k;
-         unsigned buf, bufbits= 0;
+         int buf, bufbits= 0;
          #ifndef NDEBUG
             buf= 0;
          #endif
-         for (k= digest_bits; k > 0; k-= (long)alphabet_bits) {
+         for (k= digest_bits; k > 0; k-= alphabet_bits) {
             if (bufbits < alphabet_bits) {
                /* Append the bits of another ARCFOUR output octet to <buf>. */
                ARCFOUR_STEP_3;
@@ -140,7 +141,7 @@ int main(int argc, char **argv) {
             }
             assert(bufbits >= alphabet_bits);
             {
-               int c= (int)(buf >> bufbits - alphabet_bits & alphabet_bitmask);
+               int c= buf >> bufbits - alphabet_bits & alphabet_bitmask;
                if (alphabet) c= alphabet[c];
                bufbits-= alphabet_bits;
                if (putchar(c) != c) goto wrerr;
